move functions to internal folder
